+++ x/evm/statedb/journal.go
@@ -142,6 +142,10 @@ type (
 		address *common.Address
 		slot    *common.Hash
 	}
+	transientStorageChange struct {
+		account       *common.Address
+		key, prevalue common.Hash
+	}
 )
 
 func (ch createObjectChange) Revert(s *StateDB) {
@@ -221,6 +225,14 @@ func (ch refundChange) Revert(s *StateDB) {
 }
 
 func (ch refundChange) Dirtied() *common.Address {
+	return nil
+}
+
+func (ch transientStorageChange) Revert(s *StateDB) {
+	s.setTransientState(*ch.account, ch.key, ch.prevalue)
+}
+
+func (ch transientStorageChange) Dirtied() *common.Address {
 	return nil
 }